refactor(harvester): move methods for Zenodo handling to separate file
